+++ build-docs.sh
@@ -2,4 +2,4 @@
 
 set -ev
 
-sphinx-build -n -b html -b linkcheck -d docs/_build/doctrees docs docs/_build/html 
+sphinx-build -n -b html -b linkcheck -d docs/_build/doctrees docs docs/_build/html