+++ bidict/_ordered.py
@@ -15,6 +15,8 @@ _END = _Marker('END')
 class OrderedBidictBase(BidictBase):
     """Base class for :class:`OrderedBidict` and :class:`FrozenOrderedBidict`."""
 
+    ordered = True
+
     def __init__(self, *args, **kw):
         """Like :meth:`collections.OrderedDict.__init__`."""
         # pylint: disable=super-init-not-called
@@ -125,7 +127,7 @@ class OrderedBidictBase(BidictBase):
         return key, val, isdupkey, isdupval, nodeinv, nodefwd, oldkey, oldval
 
     # pylint: disable=arguments-differ
-    def _undo_write(self, key, val, isdupkey, isdupval, nodeinv, nodefwd, oldkey, oldval):
+    def _undo_write(self, key, val, isdupkey, isdupval, nodeinv, nodefwd, oldkey, oldval):  # lgtm
         fwd = self._fwd
         inv = self._inv
         if not isdupkey and not isdupval:
@@ -149,7 +151,8 @@ class OrderedBidictBase(BidictBase):
             data.clear()
             data[key] = oldval
             data[oldval] = key
-            assert inv.pop(val) is nodefwd
+            tmp = inv.pop(val)
+            assert tmp is nodefwd
             inv[oldval] = nodefwd
             assert fwd[key] is nodefwd
         elif isdupval:
@@ -157,7 +160,8 @@ class OrderedBidictBase(BidictBase):
             data.clear()
             data[oldkey] = val
             data[val] = oldkey
-            assert fwd.pop(key) is nodeinv
+            tmp = fwd.pop(key)
+            assert tmp is nodeinv
             fwd[oldkey] = nodeinv
             assert inv[val] is nodeinv
 
@@ -168,7 +172,7 @@ class OrderedBidictBase(BidictBase):
         cur = end[_PRV if reverse else _NXT]
         while cur is not end:
             data, prv, nxt = cur
-            korv = next(iter(data))
+            korv = next(iter(data))  # lgtm [py/unguarded-next-in-generator]
             node = fwd.get(korv)
             key = korv if node is cur else data[korv]
             yield key
@@ -181,22 +185,31 @@ class OrderedBidictBase(BidictBase):
             yield key
 
     def __eq__(self, other):
+        """Like :meth:`collections.OrderedDict.__eq__`."""
         if not isinstance(other, Mapping):
             return NotImplemented
         if len(self) != len(other):
             return False
-        if self._should_compare_order_sensitive(other):
+        if self.should_compare_order_sensitive_to(other):
             return all(i == j for (i, j) in izip(iteritems(self), iteritems(other)))
         return all(self.get(k, _MISS) == v for (k, v) in iteritems(other))
 
     @staticmethod
-    def _should_compare_order_sensitive(mapping):
-        """Whether we should compare order-sensitively to ``mapping``.
+    def should_compare_order_sensitive_to(mapping):  # pylint: disable=invalid-name
+        r"""Whether we should compare order-sensitively to ``mapping``.
 
-        Returns True iff ``isinstance(mapping, OrderedBidictBase)``.
-        Override this in a subclass to customize this behavior.
+        If ``mapping`` has an ``ordered`` attribute, return its value.
+
+        Otherwise, returns True iff ``mapping`` has a ``__reversed__`` method
+        *and* ``mapping.__class__`` is not :py:class:`dict`.
+        (Unlike in CPython, in PyPY, :py:class:`dict`\s have a
+        ``__reversed__`` method but should nonetheless compare
+        order-insensitively.)
+
+        Override this method if different behavior is desired.
         """
-        return isinstance(mapping, OrderedBidictBase)
+        return getattr(mapping, 'ordered', False) or \
+            bool(getattr(mapping, '__reversed__', False)) and mapping.__class__ is not dict
 
 
 class OrderedBidict(OrderedBidictBase, bidict):