+++ tests/test_hypothesis.py
@@ -6,7 +6,7 @@ from collections import OrderedDict
 from os import getenv
 
 import pytest
-from hypothesis import assume, given, settings, unlimited
+from hypothesis import assume, given, settings
 from hypothesis.strategies import integers, lists, tuples
 from bidict import (
     IGNORE, OVERWRITE, RAISE,
@@ -15,7 +15,7 @@ from bidict import (
 from bidict.compat import iteritems
 
 
-settings.register_profile('default', settings(max_examples=200, timeout=unlimited))
+settings.register_profile('default', settings(max_examples=200, deadline=300))
 settings.load_profile(getenv('HYPOTHESIS_PROFILE', 'default'))
 
 
@@ -66,7 +66,7 @@ def test_equality(B, init):  # noqa
 @pytest.mark.parametrize('B', bidict_types)
 @given(init=inititems)
 def test_bidirectional_mappings(B, init):  # noqa
-    ordered = hasattr(B, '__reversed__')
+    ordered = bool(getattr(B, '__reversed__', None))
     C = list if ordered else sorted  # noqa
     b = B(init)
     keysf = C(k for (k, v) in iteritems(b))
@@ -99,11 +99,14 @@ def test_consistency_after_mutation(arity, methodname, B, init, arg1, arg2, item
     try:
         method(b1, *args)
     except Exception as exc:  # pylint: disable=W0703
-        # All methods should fail clean.
-        assert b1 == b0, '%r did not fail clean' % exc
-        assert b1.inv == b0.inv, '%r did not fail clean' % exc
-        return
-    # Method succeeded -> b1 should pass consistency checks.
+        # method should fail clean, i.e. b1 should be in the same state it was before the call.
+        assert b1 == b0, '%r did not fail clean: %r' % (method, exc)
+        assert b1.inv == b0.inv, '%r did not fail clean: %r' % (method, exc)
+    # Whether method failed or succeeded, b1 should pass consistency checks.
+    assert len(b1) == sum(1 for _ in iteritems(b1))
+    assert len(b1) == sum(1 for _ in iteritems(b1.inv))
+    assert b1 == dict(iteritems(b1))
+    assert b1.inv == dict(iteritems(b1.inv))
     assert b1 == to_inv_odict(iteritems(b1.inv))
     assert b1.inv == to_inv_odict(iteritems(b1))